[Aggregators]: Updated aggregators

+++ include/objectives/aggregators/AveragingAggregator.hpp
@@ -7,7 +7,7 @@ class AveragingAggregator : public AggregationFunction {
 	private:
 
 	protected:
-	float aggregateFitnesses(std::vector<float> apportionedFitnesses);
+	double aggregateFitnesses(std::vector<double> apportionedFitnesses);
 
 	public:
 };
