+++ include/objectives/aggregators/BestOfAggregator.hpp
@@ -7,7 +7,7 @@ class BestOfAggregator : public AggregationFunction {
 	private:
 
 	protected:
-	float aggregateFitnesses(std::vector<float> apportionedFitnesses);
+	double aggregateFitnesses(std::vector<double> apportionedFitnesses);
 
 	public:
 };
